requirements.txt
fastapi
uvicorn  # For running the FastAPI app
pydantic
langchain-openai
langchain-community
openai
python-dotenv  # Recommended for local environment variables